Refactor config flow tests for readability and maintainability
- Split the monolithic `test_config_flow.py` into `test_config_flow.py` and `test_options_flow.py` to better organize the tests.
- Grouped tests by feature and added BDD-style comments to improve readability.
- Simplified fixtures and reduced boilerplate by creating a `setup_test_environment` helper function.
- Improved assertions to check for specific error messages and added new test cases to increase test coverage.
